requirements.txt
shodan>=1.31.0
rich>=13.7.0
jinja2>=3.1.2
pyyaml>=6.0.1

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pre-commit>=3.5.0


